requirements.txt
langchain
openai
python-dotenv
noaa-sdk